Complete Project Management Systen with dashboard,permissions,search and DRF APIs

+++ projects/views.py
@@ -4,20 +4,26 @@ from .models import Project
 from .forms import ProjectForm
 from django.db.models import Q
 from django.db.models import Count
+from django.core.paginator import Paginator
+from django.http import HttpResponseForbidden
 
-# Create your views here.
+
 @login_required
 def project_list(request):
     projects = Project.objects.filter(
         Q(created_by=request.user)  |
         Q(members=request.user)
-    ).distinct()
+    ).distinct().order_by('-created_at')
     search=request.GET.get('search')
     if search:
         projects=projects.filter(
             Q(name__icontains=search)|
             Q(description__icontains=search)
         )
+    paginator=Paginator(projects,5)
+    page_number=request.GET.get('page')
+    projects=paginator.get_page(page_number)
+
     return render(request,'project.html',{'projects':projects})
 
 @login_required
@@ -39,6 +45,10 @@ def project_create(request):
 @login_required
 def project_update(request, pk):
     project = get_object_or_404( Project,pk=pk,created_by=request.user)
+    if request.user !=project.created_by:
+        return HttpResponseForbidden(
+            "You are not allowed to edit this project."
+        )
     if request.method == 'POST':
         form = ProjectForm(request.POST,instance=project)
         if form.is_valid():
@@ -52,6 +62,10 @@ def project_update(request, pk):
 @login_required
 def project_delete(request, pk):
     project = get_object_or_404(Project,pk=pk,created_by=request.user)
+    if request.user != project.created_by:
+        return HttpResponseForbidden(
+            "You are not allowed to delete this project."
+        )
     if request.method == 'POST':
         project.delete()
         return redirect('project_list')
@@ -61,6 +75,11 @@ def project_delete(request, pk):
 @login_required
 def project_detail(request,pk):
     project=get_object_or_404(Project,pk=pk)
+    if (request.user != project.created_by 
+        and
+        request.user not in project.members.all()
+        ):
+        return HttpResponseForbidden("You are not a member of this project.")
     tasks=project.tasks.all()
     total_tasks=tasks.count()
     completed_tasks=tasks.filter(status='Completed').count()
@@ -71,3 +90,4 @@ def project_detail(request,pk):
     
     return render(request,'project_details.html',
                   {'project':project,'tasks':tasks,'progress':progress})
+
